Renamed 8ball module
